-added Add Word view
