+++ lexicon/src/components/index.ts
@@ -1,0 +1,5 @@
+import SearchBar from './SearchBar';
+import { Tabs, TabPanel } from './Tab';
+import AddWord from './AddWord';
+
+export { SearchBar, Tabs, TabPanel, AddWord };
