+++ lexicon/src/pages/index.ts
@@ -1,0 +1,3 @@
+import Word from './Word';
+
+export { Word };
