+++ lexicon/src/types.ts
@@ -1,0 +1,4 @@
+export interface TabItem {
+  label: string;
+  value: number;
+}
